refactor code for maven project and add project pom with test dependencies
