+++ .gitignore
@@ -183,3 +183,6 @@ outg
 out/
 scores_1p
 scores_2p
+
+#Maven
+target/
