+++ src/main/java/engine/MinimalFormatter.java
